Fix external cancels propagation in rxjava2 and rxjava3 adapters
If OkHttp call is cancelled (e.g. because of a timeout) then error is
not passed to downstream resulting in a stuck rxjava invocation. Fix it
by using `disposed` instead of reusing OkHttp call cancelled state.

Separate state was added in 3f111b7b2c8c0e2ce0c4aa8f27bb94dbb6d7385f,
for the same reason but it was not used in `CallEnqueueObservable`
`onFailure` while it should be (just like it was used in try-catch in
`CallExecuteObservable`).